[FIX] code review fix
